requirements.txt
pywhatkit
twilio
google-generativeai
python-dotenv
mcp[cli]
flask
requests
pydantic
